structure/pyproject.toml
[project]
name = "gpt-cli"
version = "0.1.0"
description = "A GPT-powered CLI chatbot with TUI"
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.8"
dependencies = [
    "httpx",
    "PyYAML",
    "prompt_toolkit"
]

[tool.setuptools.packages.find]
where = ["gptcli"]
